requirements.txt
dash==2.9.3
dash[diskcache]
dash_bootstrap_components==1.0.2
dash_daq==0.5.0
dash-uploader==0.6.0
flask==3.0.0
numpy>=1.19.5
pandas==1.3.4
Pillow==8.3.2
requests==2.26.0
tiled[client]==0.1.0a107
werkzeug>=2.2.3
diskcache==5.6.3
